:recycle: Refactor ApplicationList filenames to more representative ones
