+++ src/modules/BlockchainApplication/components/AddApplication/index.js
@@ -3,10 +3,10 @@ import { translate } from 'react-i18next';
 import { View } from 'react-native';
 import HeaderBackButton from 'components/navigation/headerBackButton';
 import { useTheme } from 'contexts/ThemeContext';
-import ApplicationList from '../ApplicationList';
+import ApplicationList from '../ApplicationList/ApplicationList';
+import ApplicationRow from '../ApplicationRow/ApplicationRow';
 import getAddApplicationStyles from './styles';
 import { useApplicationsExplorer } from '../../hooks/useApplicationsExplorer';
-import BlockchainApplicationRow from '../ApplicationRow';
 
 const AddApplication = ({ navigation, t }) => {
   const { styles } = useTheme({ styles: getAddApplicationStyles() });
@@ -18,7 +18,7 @@ const AddApplication = ({ navigation, t }) => {
 
       <ApplicationList
         applications={applications}
-        Component={BlockchainApplicationRow}
+        Component={ApplicationRow}
         onItemPress={(item) =>
           navigation.navigate('ApplicationDetail', {
             chainID: item.chainID,
