+++ src/modules/BlockchainApplication/components/ApplicationList/ApplicationList.styles.js
