+++ src/modules/BlockchainApplication/components/ApplicationRow/ApplicationRow.hooks.js
