+++ src/modules/BlockchainApplication/components/ApplicationRow/ApplicationRow.styles.js
